requirements.txt
pandas
pillow
pillow-heif
geopy